refactor(storage): split chunked download + telemetry rail out of ops.py
Addresses the two sdk-review round-3 structural nits on #2479:
- storage/chunked.py now owns the resume sidecar + version-pinned chunked
  download; download_file_chunked decomposed into
  _load_and_validate_checkpoint, module-level _fetch_chunk (explicit
  params + _ChunkProgress state instead of nonlocals) and
  _handle_chunk_failure (returns restart-fresh for the first 412,
  raises typed errors otherwise). ops.py re-exports
  download_file_chunked so every existing import path keeps working
  (chunked imports ops only lazily — cycle-free).
- storage/_telemetry.py owns _classify_transfer_error,
  _record_transfer_metric, _log_transfer_progress, _throughput_mbps;
  _log_storage_event stays in ops as the per-attempt emitter.
- transfer.py: belt-and-braces sidecar unlink in the cross-store copy
  finally block (round-3 nit).
- ops.py shrinks 1596 -> 1019 lines; behavior preserved (5499 unit
  tests + load harness green).
